requirements.txt
alembic
bcrypt==4.0.1
passlib==1.7.4
fastapi
psycopg2-binary
python-jose[cryptography]
python-multipart
pydantic
pydantic[email]
pydantic-settings
python-dotenv
uvicorn[standard]
sqlalchemy